[chore] refactored List components
