+++ pages/strains/index.js
@@ -1,26 +1,81 @@
-import List, {
-  ALL_STRAINS_QUERY,
-  allStrainsQueryVars,
-} from "../../components/List";
+import List from "../../components/List";
 import { initializeApollo, addApolloState } from "../../lib/apolloClient";
 import Head from "next/head";
 import IconBubble from "../../components/IconBubble";
+import { gql, NetworkStatus, useQuery } from "@apollo/client";
+import { useEffect } from "react";
 
-const StrainsIndex = () => (
-  <>
-    <Head>
-      <title>Strains | Bubatz</title>
-      <meta
-        name="description"
-        content="Bubatz is a educational app about cannabis and everything around it."
-      />
-    </Head>
-    <List />
-    <IconBubble />
-  </>
-);
+export const ALL_STRAINS_QUERY = gql`
+  query GetAllStrains($skip: Int!, $take: Int!) {
+    allStrains(skip: $skip, take: $take) {
+      leaflyId
+      name
+      terpTop
+      thc
+      terps {
+        name
+        score
+      }
+      flowerSvg
+    }
+    getAllTerps {
+      id
+      name
+      description
+    }
+  }
+`;
+export const allStrainsQueryVars = {
+  skip: 0,
+  take: 10,
+};
 
-export async function getServerSideProps() {
+export default function StrainsPage() {
+  const { loading, error, data, fetchMore, networkStatus } = useQuery(
+    ALL_STRAINS_QUERY,
+    {
+      variables: allStrainsQueryVars,
+      notifyOnNetworkStatusChange: true,
+    }
+  );
+  const loadingMoreStrains = networkStatus === NetworkStatus.fetchMore;
+  const loadMoreStrains = () => {
+    fetchMore({
+      variables: {
+        skip: allStrains.length,
+      },
+    });
+  };
+
+  useEffect(() => {});
+
+  if (error) return <h3>Error</h3>;
+  if (loading && !loadingMoreStrains) return <h3>Loading</h3>;
+  const { allStrains } = data;
+  return (
+    <>
+      <Head>
+        <title>Strains | Bubatz</title>
+        <meta
+          name="description"
+          content="Bubatz is a educational app about cannabis and everything around it."
+        />
+      </Head>
+      <List listArr={allStrains}>
+        <button
+          className={`h-12 w-full bg-slate-200 font-medium dark:bg-slate-800 dark:text-gray-400`}
+          onClick={() => loadMoreStrains()}
+          disabled={loadingMoreStrains}
+        >
+          {loadingMoreStrains ? "Loading..." : "Show More"}
+        </button>
+      </List>
+      <IconBubble />
+    </>
+  );
+}
+
+export async function getStaticProps(context) {
   const apolloClient = initializeApollo();
 
   await apolloClient.query({
@@ -32,4 +87,3 @@ export async function getServerSideProps() {
     props: {},
   });
 }
-export default StrainsIndex;
